// src/server/types.ts
export enum RoomType {
  PRIVATE = "private",
  DEBATE = "debate",
  RANDOM = "random",
  PUBLIC = "public",
}

export enum DebateRole {
  PRO = "pro",
  CON = "con",
  MODERATOR = "moderator",
  SPECTATOR = "spectator",
}

export enum MessageType {
  TEXT = "text",
  SYSTEM = "system",
  ACTION = "action",
}